+++ NEWS.md
@@ -1,5 +1,7 @@
 # pkgload (development version)
 
+* `load_all()` now restores S3 methods registered by third party packages (#163).
+
 * `load_dll()` will now preserve the DLL name when loading instead of always using the package name. This allows packages to include DLL's with different names (#162, @dfalbel).
 
 # pkgload 1.2.1